+++ jdgfrog/app/Controller/CaseReviewsController.php
@@ -358,17 +358,18 @@ class CaseReviewsController extends AppController {
 		$this->autoRender = false;
 
 		$data = $this->DataInProgress->find('all', array('conditions' => array('DataInProgress.CaseNum' => $case_info[$index][1])));
-		$this->DataInProgress->deleteAll(array('DataInProgress.CaseNum' => $case_info[$index][1]), false);
 		$this->Datum->clear();
 
 		foreach ($data as &$d) {
 			$d['Datum'] = $d['DataInProgress'];
 			unset($d['DataInProgress']);
-
-			$this->Session->setFlash('Case Successfully Published!');
 		}		
 
-		$this->Datum->saveMany($data);
+		if ($this->Datum->saveMany($data)) {
+			$this->DataInProgress->deleteAll(array('DataInProgress.CaseNum' => $case_info[$index][1]), false);			
+			$this->Session->setFlash('Case Successfully Published!');
+
+		}
 	}
 
 	public function delete_case($CaseNum) {
